Project Files/requirements.txt
streamlit
google-generativeai
langchain
    